# python: fix enableOptimizations with clang