@mweinelt 2022-06-01
The package in this case is the cpython interpreter. You have to override it with `enableOptimizations = true` and build with clang.