@mweinelt 2022-06-01
Can confirm the PGO build using llvm11 stdenv works now.

```
# Next, run the profile task to generate the profile information.
make run_profile_task
make[1]: Entering directory '/build/Python-3.10.4'
LLVM_PROFILE_FILE="code-%p.profclangr" LD_LIBRARY_PATH=/build/Python-3.10.4 ./python -m test --pgo --timeout=1200 || true
0:00:00 load avg: 3.20 Run tests sequentially (timeout: 20 min)
0:00:00 load avg: 3.20 [ 1/44] test_array
0:00:00 load avg: 3.20 [ 2/44] test_base64
0:00:00 load avg: 3.20 [ 3/44] test_binascii -- test_base64 failed (env changed)
0:00:00 load avg: 3.20 [ 4/44] test_binop
0:00:00 load avg: 3.20 [ 5/44] test_bisect
0:00:00 load avg: 3.20 [ 6/44] test_bytes
0:00:02 load avg: 3.20 [ 7/44] test_bz2 -- test_bytes failed (env changed)
0:00:03 load avg: 3.20 [ 8/44] test_cmath
0:00:03 load avg: 3.20 [ 9/44] test_codecs
0:00:04 load avg: 3.03 [10/44] test_collections
0:00:05 load avg: 3.03 [11/44] test_complex
0:00:05 load avg: 3.03 [12/44] test_dataclasses
0:00:05 load avg: 3.03 [13/44] test_datetime
0:00:08 load avg: 3.03 [14/44] test_decimal
0:00:11 load avg: 2.86 [15/44] test_difflib
0:00:11 load avg: 2.86 [16/44] test_embed
test test_embed failed
0:00:15 load avg: 2.79 [17/44] test_float -- test_embed failed (4 failures)
0:00:15 load avg: 2.79 [18/44] test_fstring
0:00:15 load avg: 2.79 [19/44] test_functools
0:00:16 load avg: 2.79 [20/44] test_generators
0:00:16 load avg: 2.79 [21/44] test_hashlib
0:00:16 load avg: 2.79 [22/44] test_heapq
0:00:17 load avg: 2.79 [23/44] test_int
0:00:17 load avg: 2.79 [24/44] test_itertools
0:00:18 load avg: 2.79 [25/44] test_json
0:00:20 load avg: 2.73 [26/44] test_long -- test_json failed (env changed)
0:00:22 load avg: 2.73 [27/44] test_lzma
0:00:22 load avg: 2.73 [28/44] test_math
0:00:24 load avg: 2.59 [29/44] test_memoryview
0:00:24 load avg: 2.59 [30/44] test_operator
0:00:24 load avg: 2.59 [31/44] test_ordered_dict
0:00:25 load avg: 2.59 [32/44] test_patma
0:00:25 load avg: 2.59 [33/44] test_pickle
0:00:29 load avg: 2.54 [34/44] test_pprint
0:00:29 load avg: 2.54 [35/44] test_re
0:00:29 load avg: 2.54 [36/44] test_set
0:00:32 load avg: 2.54 [37/44] test_sqlite
0:00:34 load avg: 2.54 [38/44] test_statistics -- test_sqlite failed (env changed)
0:00:35 load avg: 2.50 [39/44] test_struct
0:00:35 load avg: 2.50 [40/44] test_tabnanny -- test_struct failed (env changed)
0:00:35 load avg: 2.50 [41/44] test_time -- test_tabnanny failed (env changed)
0:00:38 load avg: 2.50 [42/44] test_unicode
0:00:39 load avg: 2.38 [43/44] test_xml_etree -- test_unicode failed (env changed)
0:00:39 load avg: 2.38 [44/44] test_xml_etree_c

Total duration: 40.0 sec
Tests result: FAILURE
make[1]: Leaving directory '/build/Python-3.10.4'
make build_all_merge_profile
make[1]: Entering directory '/build/Python-3.10.4'
/nix/store/1zy8n51d4i0nfyxnb9ramwcmrd7mjz45-llvm-11.1.0/bin/llvm-profdata merge -output=code.profclangd *.profclangr
make[1]: Leaving directory '/build/Python-3.10.4'
# Remove profile generation binary since we are done with it.
make clean-retain-profile
```

Benchmarking shows this is not really significant, maybe I'm holding things wrong. I did `python3.override { stdenv = clangStdenv; enableOptimizations = true; }`.

```
### 2to3 ###
Mean +- std dev: 466 ms +- 13 ms -> 461 ms +- 8 ms: 1.01x faster
Not significant

### chameleon ###
Mean +- std dev: 11.8 ms +- 0.3 ms -> 11.7 ms +- 0.3 ms: 1.01x faster
Not significant

### chaos ###
Mean +- std dev: 148 ms +- 3 ms -> 147 ms +- 2 ms: 1.01x faster
Not significant

### crypto_pyaes ###
Mean +- std dev: 148 ms +- 3 ms -> 148 ms +- 3 ms: 1.00x slower
Not significant

### deltablue ###
Mean +- std dev: 10.4 ms +- 0.1 ms -> 10.4 ms +- 0.2 ms: 1.00x slower
Not significant

### django_template ###
Mean +- std dev: 69.9 ms +- 1.2 ms -> 71.3 ms +- 1.4 ms: 1.02x slower
Not significant

### dulwich_log ###
Mean +- std dev: 117 ms +- 3 ms -> 116 ms +- 1 ms: 1.01x faster
Not significant

### fannkuch ###
Mean +- std dev: 594 ms +- 6 ms -> 593 ms +- 8 ms: 1.00x faster
Not significant

### float ###
Mean +- std dev: 138 ms +- 3 ms -> 138 ms +- 2 ms: 1.00x faster
Not significant

### genshi_text ###
Mean +- std dev: 40.0 ms +- 1.1 ms -> 40.6 ms +- 3.2 ms: 1.01x slower
Not significant

### genshi_xml ###
Mean +- std dev: 89.7 ms +- 4.1 ms -> 90.0 ms +- 3.4 ms: 1.00x slower
Not significant

### go ###
Mean +- std dev: 297 ms +- 7 ms -> 300 ms +- 10 ms: 1.01x slower
Not significant

### hexiom ###
Mean +- std dev: 12.2 ms +- 0.1 ms -> 12.3 ms +- 0.1 ms: 1.01x slower
Not significant

### html5lib ###
Mean +- std dev: 114 ms +- 4 ms -> 113 ms +- 5 ms: 1.01x faster
Not significant

### json_dumps ###
Mean +- std dev: 16.7 ms +- 0.4 ms -> 16.5 ms +- 0.3 ms: 1.01x faster
Not significant

### json_loads ###
Mean +- std dev: 31.2 us +- 1.4 us -> 31.1 us +- 0.7 us: 1.00x faster
Not significant

### logging_format ###
Mean +- std dev: 14.8 us +- 0.2 us -> 14.6 us +- 0.3 us: 1.01x faster
Not significant

### logging_silent ###
Mean +- std dev: 245 ns +- 8 ns -> 246 ns +- 7 ns: 1.00x slower
Not significant

### logging_simple ###
Mean +- std dev: 13.1 us +- 0.3 us -> 13.1 us +- 0.3 us: 1.00x faster
Not significant

### mako ###
Mean +- std dev: 19.5 ms +- 0.3 ms -> 19.4 ms +- 0.3 ms: 1.00x faster
Not significant

### meteor_contest ###
Mean +- std dev: 131 ms +- 5 ms -> 128 ms +- 2 ms: 1.03x faster
Significant (t=5.60)

### nbody ###
Mean +- std dev: 180 ms +- 6 ms -> 179 ms +- 3 ms: 1.01x faster
Not significant

### nqueens ###
Mean +- std dev: 136 ms +- 3 ms -> 136 ms +- 2 ms: 1.00x slower
Not significant

### pathlib ###
Mean +- std dev: 26.2 ms +- 0.7 ms -> 26.2 ms +- 1.1 ms: 1.00x faster
Not significant

### pickle ###
Mean +- std dev: 13.3 us +- 0.3 us -> 13.2 us +- 0.3 us: 1.01x faster
Not significant

### pickle_dict ###
Mean +- std dev: 30.5 us +- 1.3 us -> 29.8 us +- 0.6 us: 1.02x faster
Significant (t=3.50)

### pickle_list ###
Mean +- std dev: 4.80 us +- 0.07 us -> 4.79 us +- 0.12 us: 1.00x faster
Not significant

### pickle_pure_python ###
Mean +- std dev: 618 us +- 17 us -> 616 us +- 18 us: 1.00x faster
Not significant

### pidigits ###
Mean +- std dev: 216 ms +- 2 ms -> 216 ms +- 3 ms: 1.00x faster
Not significant

### pyflate ###
Mean +- std dev: 837 ms +- 22 ms -> 828 ms +- 11 ms: 1.01x faster
Not significant

### python_startup ###
Mean +- std dev: 12.2 ms +- 0.6 ms -> 12.2 ms +- 0.3 ms: 1.00x faster
Not significant

### python_startup_no_site ###
Mean +- std dev: 8.11 ms +- 0.18 ms -> 8.12 ms +- 0.27 ms: 1.00x slower
Not significant

### raytrace ###
Mean +- std dev: 696 ms +- 10 ms -> 702 ms +- 15 ms: 1.01x slower
Not significant

### regex_compile ###
Mean +- std dev: 236 ms +- 10 ms -> 238 ms +- 11 ms: 1.01x slower
Not significant

### regex_dna ###
Mean +- std dev: 199 ms +- 8 ms -> 198 ms +- 4 ms: 1.01x faster
Not significant

### regex_effbot ###
Mean +- std dev: 3.39 ms +- 0.05 ms -> 3.40 ms +- 0.05 ms: 1.00x slower
Not significant

### regex_v8 ###
Mean +- std dev: 28.0 ms +- 0.7 ms -> 27.9 ms +- 0.3 ms: 1.01x faster
Not significant

### richards ###
Mean +- std dev: 108 ms +- 3 ms -> 109 ms +- 4 ms: 1.01x slower
Not significant

### scimark_fft ###
Mean +- std dev: 491 ms +- 10 ms -> 498 ms +- 12 ms: 1.01x slower
Not significant

### scimark_lu ###
Mean +- std dev: 212 ms +- 4 ms -> 212 ms +- 4 ms: 1.00x faster
Not significant

### scimark_monte_carlo ###
Mean +- std dev: 134 ms +- 2 ms -> 133 ms +- 3 ms: 1.01x faster
Not significant

### scimark_sor ###
Mean +- std dev: 254 ms +- 6 ms -> 253 ms +- 5 ms: 1.01x faster
Not significant

### scimark_sparse_mat_mult ###
Mean +- std dev: 6.82 ms +- 0.07 ms -> 6.98 ms +- 0.43 ms: 1.02x slower
Significant (t=-2.86)

### spectral_norm ###
Mean +- std dev: 186 ms +- 5 ms -> 187 ms +- 8 ms: 1.00x slower
Not significant

### sqlalchemy_declarative ###
Mean +- std dev: 201 ms +- 8 ms -> 198 ms +- 6 ms: 1.01x faster
Not significant

### sqlalchemy_imperative ###
Mean +- std dev: 32.7 ms +- 1.1 ms -> 32.2 ms +- 1.7 ms: 1.02x faster
Not significant

### sqlite_synth ###
Mean +- std dev: 3.40 us +- 0.08 us -> 3.41 us +- 0.15 us: 1.00x slower
Not significant

### sympy_expand ###
Mean +- std dev: 773 ms +- 16 ms -> 772 ms +- 18 ms: 1.00x faster
Not significant

### sympy_integrate ###
Mean +- std dev: 31.6 ms +- 0.7 ms -> 31.4 ms +- 0.7 ms: 1.01x faster
Not significant

### sympy_str ###
Mean +- std dev: 459 ms +- 11 ms -> 455 ms +- 8 ms: 1.01x faster
Not significant

### sympy_sum ###
Mean +- std dev: 248 ms +- 3 ms -> 250 ms +- 8 ms: 1.01x slower
Not significant

### telco ###
Mean +- std dev: 8.68 ms +- 0.20 ms -> 8.70 ms +- 0.13 ms: 1.00x slower
Not significant

### tornado_http ###
Mean +- std dev: 201 ms +- 6 ms -> 198 ms +- 7 ms: 1.02x faster
Not significant

### unpack_sequence ###
Mean +- std dev: 58.4 ns +- 1.3 ns -> 58.1 ns +- 0.9 ns: 1.00x faster
Not significant

### unpickle ###
Mean +- std dev: 17.0 us +- 0.7 us -> 17.2 us +- 1.0 us: 1.01x slower
Not significant

### unpickle_list ###
Mean +- std dev: 5.60 us +- 0.21 us -> 5.54 us +- 0.15 us: 1.01x faster
Not significant

### unpickle_pure_python ###
Mean +- std dev: 420 us +- 10 us -> 420 us +- 13 us: 1.00x faster
Not significant

### xml_etree_generate ###
Mean +- std dev: 121 ms +- 2 ms -> 122 ms +- 5 ms: 1.01x slower
Not significant

### xml_etree_iterparse ###
Mean +- std dev: 127 ms +- 3 ms -> 127 ms +- 4 ms: 1.00x slower
Not significant

### xml_etree_parse ###
Mean +- std dev: 173 ms +- 3 ms -> 176 ms +- 8 ms: 1.02x slower
Not significant

### xml_etree_process ###
Mean +- std dev: 99.9 ms +- 1.4 ms -> 99.9 ms +- 1.6 ms: 1.00x faster
Not significant